refactor(fix): changes to address review comments
