+++ packages/opencensus-core/src/exporters/types.ts
@@ -18,7 +18,7 @@ import {Measurement, View} from '../stats/types';
 import * as configTypes from '../trace/config/types';
 import * as modelTypes from '../trace/model/types';
 
-/** Defines an exporter interface. */
+/** Defines a trace exporter interface. */
 export interface Exporter extends modelTypes.SpanEventListener {
   /**
    * Sends a list of root spans to the service.
@@ -27,9 +27,21 @@ export interface Exporter extends modelTypes.SpanEventListener {
   publish(rootSpans: modelTypes.RootSpan[]): Promise<number|string|void>;
 }
 
-/** Defines a StatsEventListener interface */
+/**
+ * An interface that describes the possible events that will be emitted from a
+ * Stats instance. Stats exporters should implement this interface.
+ */
 export interface StatsEventListener {
+  /**
+   * Is called whenever a new view is registered
+   * @param view The registered view
+   */
   onRegisterView(view: View): void;
+  /**
+   * Is called whenever a new measurement is recorded.
+   * @param view The view related to the measurement
+   * @param measurement The recorded measurement
+   */
   onRecord(view: View, measurement: Measurement): void;
 }
